move check provisioned into util function

+++ public/app/api/clients/folder/v1beta1/hooks.ts
@@ -10,7 +10,6 @@ import {
 } from 'app/features/browse-dashboards/api/browseDashboardsAPI';
 import { FolderDTO } from 'app/types/folders';
 
-import appEvents from '../../../../core/app_events';
 import kbn from '../../../../core/utils/kbn';
 import {
   AnnoKeyCreatedBy,
@@ -21,21 +20,16 @@ import {
   DeprecatedInternalId,
   ManagerKind,
 } from '../../../../features/apiserver/types';
-import { isProvisionedFolder } from '../../../../features/browse-dashboards/api/isProvisioned';
 import { PAGE_SIZE } from '../../../../features/browse-dashboards/api/services';
 import { refetchChildren, refreshParents } from '../../../../features/browse-dashboards/state/actions';
 import { GENERAL_FOLDER_UID } from '../../../../features/search/constants';
 import { useDispatch } from '../../../../types/store';
 import { useGetDisplayMappingQuery } from '../../iam/v0alpha1';
 
+import { isProvisionedFolderCheck } from './utils';
 import { rootFolder, sharedWithMeFolder } from './virtualFolders';
 
-import {
-  useGetFolderQuery,
-  useGetFolderParentsQuery,
-  useDeleteFolderMutation,
-  folderAPIv1beta1 as folderAPI,
-} from './index';
+import { useGetFolderQuery, useGetFolderParentsQuery, useDeleteFolderMutation } from './index';
 
 function getFolderUrl(uid: string, title: string): string {
   // mimics https://github.com/grafana/grafana/blob/79fe8a9902335c7a28af30e467b904a4ccfac503/pkg/services/dashboards/models.go#L188
@@ -211,21 +205,9 @@ export function useDeleteMultipleFoldersMutationFacade() {
     // Delete all the folders sequentially
     // TODO error handling here
     for (const folderUID of folderUIDs) {
-      if (config.featureToggles.provisioning) {
-        const folder = await dispatch(folderAPI.endpoints.getFolder.initiate({ name: folderUID }));
-        // TODO: taken from browseDashboardAPI as it is, but this error handling should be moved up to UI code.
-        if (folder.data && isProvisionedFolder(folder.data)) {
-          appEvents.publish({
-            type: AppEvents.alertWarning.name,
-            payload: [
-              t(
-                'folders.api.folder-delete-error-provisioned',
-                'Cannot delete provisioned folder. To remove it, delete it from the repository and synchronise to apply the changes.'
-              ),
-            ],
-          });
-          continue;
-        }
+      // This also shows warning alert
+      if (await isProvisionedFolderCheck(dispatch, folderUID)) {
+        continue;
       }
       const result = await deleteFolder({ name: folderUID });
       if (!result.error) {
